Implemented app_node_cmd project type and tests
replaced tabs with spaces on all project files
adapted this project to be an app_node_cmd one
improved turbobuilder json schema

+++ TurboBuilder-Node/src/main/js/clean.js
@@ -21,6 +21,11 @@ exports.execute = function (alsoCleanSync = false) {
     
     console.log("\nclean start");
     
+    if(global.setup.build.app_node_cmd){
+        
+        return console.success('Clean is not necessary for node cmd apps');
+    }
+    
     if(fm.isDirectory(global.runtimePaths.target) &&
             !fm.deleteDirectory(global.runtimePaths.target) &&
             !fm.isDirectoryEmpty(global.runtimePaths.target)){
